Renamed the package listing, added a pie chart.
You can also now specify a "function" that will be called to process the collection. The comments have been updated to correctly identify the additional options.

The next step is to create some default handlers for common charts. These will include the things necesary for each kind of chart (such as the jqplot renderer or what not).